Create l_test_human_input.py
This file will be used as the first part of my attempts to reduce training time, where I will train the AI on human generated inputs. I will then finish training without human input and compare it to an algorithm that has trained for the same total number of epochs.